Make PersonDetailProvider a bit more obvious.
